% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrap_coefficients.R
\name{plot_bootstrap_coefficients}
\alias{plot_bootstrap_coefficients}
\title{Plot Bootstrap LASSO Coefficient Stability}
\usage{
plot_bootstrap_coefficients(
  coef_matrix,
  dot_col = "lightblue",
  top_n = 20,
  return_data = FALSE,
  n_boot = 500
)
}
\arguments{
\item{coef_matrix}{The coefficient matrix which produced by \code{train_temporal_model} function.}

\item{dot_col}{The point colors of the dot plot. Default is lightblue.}

\item{top_n}{Integer, number of top features to display. Default is 20.}

\item{return_data}{Logical, whether to return the summary table. Default is FALSE.}

\item{n_boot}{Bootstrapping numbers. Default is 500.}
}
\value{
A \code{ggplot2} dot plot showing selection frequency and mean coefficients.
If \code{return = TRUE}, returns a summary data frame instead.
}
\description{
This function visulizas the frequency and average effect size of gene selected in an ensemble of bootstrapped LASSO models.
}
\examples{
# boot_models <- bootstrap_temporal_models(...)
# plot_bootstrap_coefficients(boot_models$models)
}
