% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_temporal_gap.R
\name{predict_temporal_gap}
\alias{predict_temporal_gap}
\title{Predict the disease process time and Compute Temporal Gap
This function uses a trained LASSO model to predict disease process time for gene expression data, and
optionally calculates the Temporal Gap between predicted and actual post-injury hours.}
\usage{
predict_temporal_gap(
  model,
  expression_matrix,
  actual_time = NULL,
  n_boot = 500
)
}
\arguments{
\item{model}{A trained \code{cv.glmnet} object return from \code{trian_temporal_model()}}

\item{expression_matrix}{A numeric matrix (features x samples) of gene/protein/lipid... expression values for prediction.}

\item{actual_time}{Optional. A numeric vector of actual post-injury hours (for computing TempoGap).}

\item{nboot}{Bootstrapping numbers. Default is 500.}
}
\value{
A data frame with predicted time, actual time (if provided), and TempoGap (if applicable).
}
\description{
Predict the disease process time and Compute Temporal Gap
This function uses a trained LASSO model to predict disease process time for gene expression data, and
optionally calculates the Temporal Gap between predicted and actual post-injury hours.
}
\examples{
# set.seed(123)
# expr_train <- matrix(rnorm(1000), nrow = 100, ncol = 10)
# time_train <- runif(10, 0, 72)
# model <- train_temporal_model(expr_train, time_train)$model

# expr_test <- matrix(rnorm(500), nrow = 100, ncol = 5)
# time_test <- runif(5, 0, 72)
# results <- predict_temporal_model(model, expr_test, time_test)
}
