% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_temporal_model.R
\name{train_temporal_model}
\alias{train_temporal_model}
\title{Train Temporal Model
Purpose: Trains the LASSO mdoel to predict recovery class of disease based on expression data}
\usage{
train_temporal_model(
  expression_matrix,
  time_vector,
  cv_folds = 5,
  alpha = 1,
  seed = NULL,
  n_boot = 500
)
}
\arguments{
\item{expression_matrix}{A numeric matrix (features x samples) of gene/protein/lipid... expression values.}

\item{cv_folds}{Integer, number of folds for cross-validation. Default is 5.}

\item{alpha}{Elastic net mixing parameter. Default is 1 (LASSO). If alpha is a vector of continuous numbers, hyperparameter optimization of alpha will be performed based on the vector.}

\item{seed}{Optional, a random seed for reproducibility.}

\item{time_vextor}{A numeric vextor of disease process timing data for each sample. Note: It must have same units.}

\item{nboot}{Bootstrapping numbers. Default is 500.}
}
\description{
Train Temporal Model
Purpose: Trains the LASSO mdoel to predict recovery class of disease based on expression data
}
\examples{
# set.seed(42)
# data <- matrix(rnorm(1000), nrow = 100, ncol = 10)
# time <- runif(10, 0, 72)
# model <- train_temporal_model(data, time)
}
